Added build_choose_player
Also added quit_action and back_action in relavent places.
There is some major code duplication here, and there will be much more if we continue this way...
Must add titles. Note: Titles can be depended on game_state fields that are not yet defined, like if we set for cat or mouse.